+++ packages/dart/noports_core/lib/src/srv/srv_impl.dart
@@ -37,7 +37,7 @@ class SrvImplExec implements Srv<Process> {
   final bool? bindLocalPort;
 
   @override
-  final String? rvdAuthString;
+  final RelayAuthenticator? relayAuthenticator;
 
   @override
   final String? sessionAESKeyString;
@@ -60,13 +60,14 @@ class SrvImplExec implements Srv<Process> {
     this.localPort,
     this.localHost,
     this.bindLocalPort = false,
-    this.rvdAuthString,
+    required this.relayAuthenticator,
     this.sessionAESKeyString,
     this.sessionIVString,
     required this.multi,
     required this.timeout,
     this.controlChannelHeartbeat,
   }) {
+    logger.info('New SrvImplDart - localPort $localPort - timeout $timeout');
     if (localPort == null) {
       throw ArgumentError('localPort must be non-null');
     }
@@ -111,9 +112,12 @@ class SrvImplExec implements Srv<Process> {
       rvArgs.add('--bind-local-port');
     }
     Map<String, String> environment = {};
-    if (rvdAuthString != null) {
-      rvArgs.addAll(['--rv-auth']);
-      environment['RV_AUTH'] = rvdAuthString!;
+
+    if (relayAuthenticator != null) {
+      rvArgs.addAll(relayAuthenticator!.rvArgs);
+      for (final String name in relayAuthenticator!.envMap.keys) {
+        environment[name] = relayAuthenticator!.envMap[name]!;
+      }
     }
     if (sessionAESKeyString != null && sessionIVString != null) {
       rvArgs.addAll(['--rv-e2ee']);
@@ -190,7 +194,7 @@ class SrvImplInline implements Srv<SSHSocket> {
   final String? localHost = null;
 
   @override
-  final String? rvdAuthString;
+  final RelayAuthenticator? relayAuthenticator;
 
   @override
   final String? sessionAESKeyString;
@@ -210,13 +214,14 @@ class SrvImplInline implements Srv<SSHSocket> {
   SrvImplInline(
     this.streamingHost,
     this.streamingPort, {
-    this.rvdAuthString,
+    required this.relayAuthenticator,
     this.sessionAESKeyString,
     this.sessionIVString,
     this.multi = false,
     required this.timeout,
     required this.controlChannelHeartbeat,
   }) {
+    logger.info('New SrvImplInline - timeout $timeout');
     if ((sessionAESKeyString == null && sessionIVString != null) ||
         (sessionAESKeyString != null && sessionIVString == null)) {
       throw ArgumentError('Both AES key and IV are required, or neither');
@@ -259,15 +264,28 @@ class SrvImplInline implements Srv<SSHSocket> {
           ' at $streamingHost:$streamingPort');
       Socket socket = await Socket.connect(streamingHost, streamingPort);
 
-      // Authenticate if we have an rvdAuthString
-      if (rvdAuthString != null) {
+      // Authenticate if we have a relayAuthenticator
+      Stream<Uint8List>? socketStream;
+      if (relayAuthenticator != null) {
+        bool authenticated;
         logger.info('run() authenticating to rvd');
-        socket.writeln(rvdAuthString);
-        await socket.flush();
+        (authenticated, socketStream) =
+            await relayAuthenticator!.authenticate(socket);
+        if (!authenticated) {
+          throw Exception('Authentication failed');
+        }
+      } else {
+        socketStream = socket;
       }
 
-      WrappedSSHSocket sshSocket =
-          WrappedSSHSocket(socket, rvdAuthString, encrypter, decrypter);
+      WrappedSSHSocket sshSocket = WrappedSSHSocket(
+        socketStream!,
+        socket,
+        encrypter,
+        decrypter,
+        onClose: () async => socket.close(),
+        onDestroy: () => socket.destroy(),
+      );
 
       return sshSocket;
     } catch (e) {
@@ -281,45 +299,53 @@ class SrvImplInline implements Srv<SSHSocket> {
 /// - Wrap the StreamSink with encrypter
 /// - Wrap the Stream with decrypter
 class WrappedSSHSocket implements SSHSocket {
-  /// The actual underlying socket
-  final Socket socket;
-  final String? rvdAuthString;
+  final Stream<Uint8List> underlyingStream;
+  final IOSink underlyingSink;
   final DataTransformer? encrypter;
   final DataTransformer? decrypter;
 
   late StreamSink<List<int>> _sink;
   late Stream<Uint8List> _stream;
 
+  Future<void> Function() onClose;
+  void Function() onDestroy;
+
   WrappedSSHSocket(
-      this.socket, this.rvdAuthString, this.encrypter, this.decrypter) {
+    this.underlyingStream,
+    this.underlyingSink,
+    this.encrypter,
+    this.decrypter, {
+    required this.onClose,
+    required this.onDestroy,
+  }) {
     if (encrypter == null) {
-      _sink = socket;
+      _sink = underlyingSink;
     } else {
       StreamController<Uint8List> sc = StreamController<Uint8List>();
       Stream<List<int>> encrypted = encrypter!(sc.stream);
-      encrypted.listen(socket.add);
+      encrypted.listen(underlyingSink.add);
       _sink = sc;
     }
 
     if (decrypter == null) {
-      _stream = socket;
+      _stream = underlyingStream;
     } else {
-      _stream = decrypter!(socket).cast<Uint8List>();
+      _stream = decrypter!(underlyingStream).cast<Uint8List>();
     }
   }
 
   @override
   Future<void> close() async {
-    await socket.close();
+    await onClose();
   }
 
   @override
   void destroy() {
-    socket.destroy();
+    onDestroy();
   }
 
   @override
-  Future<void> get done => socket.done;
+  Future<void> get done => sink.done;
 
   @override
   StreamSink<List<int>> get sink => _sink;
@@ -346,7 +372,7 @@ class SrvImplDart implements Srv<SocketConnector> {
   final bool bindLocalPort;
 
   @override
-  final String? rvdAuthString;
+  final RelayAuthenticator? relayAuthenticator;
 
   @override
   final String? sessionAESKeyString;
@@ -373,7 +399,7 @@ class SrvImplDart implements Srv<SocketConnector> {
     required this.localPort,
     required this.bindLocalPort,
     this.localHost,
-    this.rvdAuthString,
+    required this.relayAuthenticator,
     this.sessionAESKeyString,
     this.sessionIVString,
     this.multi = false,
@@ -381,7 +407,7 @@ class SrvImplDart implements Srv<SocketConnector> {
     required this.timeout,
     this.controlChannelHeartbeat,
   }) {
-    logger.info('New SrvImplDart - localPort $localPort');
+    logger.info('New SrvImplDart - localPort $localPort - timeout $timeout');
     if ((sessionAESKeyString == null && sessionIVString != null) ||
         (sessionAESKeyString != null && sessionIVString == null)) {
       throw ArgumentError('Both AES key and IV are required, or neither');
@@ -506,13 +532,25 @@ class SrvImplDart implements Srv<SocketConnector> {
       transformBtoA: decrypter,
       multi: multi,
       timeout: timeout,
-      beforeJoining: (Side sideA, Side sideB) {
+      beforeJoining: (Side sideA, Side sideB) async {
         logger.info('beforeJoining called');
         // Authenticate the sideB socket (to the rvd)
-        if (rvdAuthString != null) {
+        if (relayAuthenticator != null) {
           logger.info('_runClientSideSingle authenticating'
               ' new connection to rvd');
-          sideB.socket.writeln(rvdAuthString);
+          try {
+            var (authenticated, authenticatedStream) =
+                await relayAuthenticator!.authenticate(sideB.socket);
+            if (!authenticated || authenticatedStream == null) {
+              sideB.socket.destroy();
+            } else {
+              sideB.stream = authenticatedStream;
+            }
+          } catch (err) {
+            logger.severe('_runClientSideSingle'
+                ' Failed to authenticate to relay: $err');
+            sideB.socket.destroy();
+          }
         }
       },
     );
@@ -529,22 +567,49 @@ class SrvImplDart implements Srv<SocketConnector> {
         streamingHost, streamingPort,
         timeout: Duration(seconds: 10));
     // Authenticate the control channel
-    if (rvdAuthString != null) {
+    Stream<Uint8List>? authenticatedControlSocketStream;
+    if (relayAuthenticator != null) {
       logger.info('_runClientSideMulti authenticating'
           ' control channel connection to rvd');
-      sessionControlSocket.writeln(rvdAuthString);
+      bool authenticated;
+      try {
+        (authenticated, authenticatedControlSocketStream) =
+            await relayAuthenticator!.authenticate(sessionControlSocket);
+        if (!authenticated || authenticatedControlSocketStream == null) {
+          sessionControlSocket.destroy();
+          throw Exception('_runClientSideMulti'
+              ' Failed to authenticate control socket');
+        }
+      } catch (err) {
+        logger.severe('_runClientSideMulti'
+            ' Failed to authenticate control socket: $err');
+        sessionControlSocket.destroy();
+        rethrow;
+      }
+    } else {
+      authenticatedControlSocketStream = sessionControlSocket;
     }
 
     if (sessionAESKeyString != null && sessionIVString != null) {
-      logger
-          .info('_runClientSideMulti: On the client-side traffic is encrypted');
+      logger.info('_runClientSideMulti:'
+          ' On the client-side traffic is encrypted');
       socketConnector = await _clientSideEncryptedSocket(
-          sessionControlSocket, socketConnector, relayAddress, timeout);
+        authenticatedControlSocketStream,
+        sessionControlSocket,
+        socketConnector,
+        relayAddress,
+        timeout,
+      );
     } else {
-      logger.info(
-          '_runClientSideMulti: On the client-side traffic is transmitted in plain text');
+      logger.info('_runClientSideMulti:'
+          ' On the client-side traffic is transmitted in plain text');
       socketConnector = await _clientSidePlainSocket(
-          sessionControlSocket, socketConnector, relayAddress, timeout);
+        authenticatedControlSocketStream,
+        sessionControlSocket,
+        socketConnector,
+        relayAddress,
+        timeout,
+      );
     }
 
     logger.info('_runClientSideMulti serverToSocket is ready');
@@ -558,19 +623,22 @@ class SrvImplDart implements Srv<SocketConnector> {
 
   /// On the client side, the data in this socket remains unencrypted and is transmitted in plain text
   Future<SocketConnector> _clientSidePlainSocket(
-      Socket sessionControlSocket,
+      Stream<Uint8List> sessionControlSocketStream,
+      IOSink sessionControlSocketSink,
       SocketConnector? socketConnector,
       InternetAddress relayAddress,
       Duration timeout) async {
-    sessionControlSocket.listen((event) {
+    sessionControlSocketStream.listen((event) {
       String response = String.fromCharCodes(event).trim();
-      logger.info('_runClientSideMulti'
+      logger.info('_runClientSideMulti (_client_sidePlainSocket)'
           ' Received control channel response: [$response]');
     }, onError: (e) {
-      logger.severe('_runClientSideMulti controlSocket error: $e');
+      logger.severe('_runClientSideMulti (_clientSidePlainSocket)'
+          ' controlSocket error: $e');
       socketConnector?.close();
     }, onDone: () {
-      logger.info('_runClientSideMulti controlSocket done');
+      logger.info('_runClientSideMulti (_clientSidePlainSocket)'
+          ' controlSocket done');
       socketConnector?.close();
     });
     socketConnector = await SocketConnector.serverToSocket(
@@ -581,16 +649,28 @@ class SrvImplDart implements Srv<SocketConnector> {
       logger: ioSinkForLogger(logger),
       multi: multi,
       timeout: timeout,
-      beforeJoining: (Side sideA, Side sideB) {
-        logger.info('_runClientSideMulti Sending connect request');
-        sessionControlSocket.add(
-          Uint8List.fromList('connect:no:encrypt\n'.codeUnits),
-        );
+      beforeJoining: (Side sideA, Side sideB) async {
+        logger.info('_runClientSideMulti (_clientSidePlainSocket)'
+            ' Sending connect request');
+        sessionControlSocketSink
+            .add(Uint8List.fromList('connect:no:encrypt\n'.codeUnits));
         // Authenticate the sideB socket (to the rvd)
-        if (rvdAuthString != null) {
-          logger
-              .info('_runClientSideMulti authenticating new connection to rvd');
-          sideB.socket.writeln(rvdAuthString);
+        if (relayAuthenticator != null) {
+          logger.info('_runClientSideMulti (_clientSidePlainSocket)'
+              'authenticating new connection to rvd');
+          try {
+            var (authenticated, authenticatedStream) =
+                await relayAuthenticator!.authenticate(sideB.socket);
+            if (!authenticated || authenticatedStream == null) {
+              sideB.socket.destroy();
+            } else {
+              sideB.stream = authenticatedStream;
+            }
+          } catch (err) {
+            logger.severe('_runClientSideMulti (_clientSidePlainSocket)'
+                ' Failed to authenticate to relay: $err');
+            sideB.socket.destroy();
+          }
         }
       },
     );
@@ -599,7 +679,8 @@ class SrvImplDart implements Srv<SocketConnector> {
 
   /// On the client side, the data in encrypted and is transmitted through this socket.
   Future<SocketConnector> _clientSideEncryptedSocket(
-      Socket sessionControlSocket,
+      Stream<Uint8List> sessionControlSocketStream,
+      IOSink sessionControlSocketSink,
       SocketConnector? socketConnector,
       InternetAddress relayAddress,
       Duration timeout) async {
@@ -610,19 +691,22 @@ class SrvImplDart implements Srv<SocketConnector> {
 
     // Listen to stream which is decrypting the socket stream
     // Write to a stream controller which encrypts and writes to the socket
-    Stream<List<int>> controlStream = controlDecrypter(sessionControlSocket);
+    Stream<List<int>> controlStream =
+        controlDecrypter(sessionControlSocketStream);
     StreamController<Uint8List> controlSink = StreamController<Uint8List>();
-    controlEncrypter(controlSink.stream).listen(sessionControlSocket.add);
+    controlEncrypter(controlSink.stream).listen(sessionControlSocketSink.add);
 
     controlStream.listen((event) {
       String response = String.fromCharCodes(event).trim();
-      logger.info('_runClientSideMulti'
+      logger.info('_runClientSideMulti (_clientSideEncryptedSocket)'
           ' Received control channel response: [$response]');
     }, onError: (e) {
-      logger.severe('_runClientSideMulti controlSocket error: $e');
+      logger.severe('_runClientSideMulti  (_clientSideEncryptedSocket)'
+          ' controlSocket error: $e');
       socketConnector?.close();
     }, onDone: () {
-      logger.info('_runClientSideMulti controlSocket done');
+      logger.info('_runClientSideMulti  (_clientSideEncryptedSocket)'
+          ' controlSocket done');
       socketConnector?.close();
     });
 
@@ -648,6 +732,8 @@ class SrvImplDart implements Srv<SocketConnector> {
     }
 
     logger.info('_runClientSideMulti calling SocketConnector.serverToSocket');
+    logger.info('_runClientSideMulti  (_clientSideEncryptedSocket)'
+        ' calling SocketConnector.serverToSocket');
     socketConnector = await SocketConnector.serverToSocket(
       portA: localPort,
       addressB: relayAddress,
@@ -663,16 +749,33 @@ class SrvImplDart implements Srv<SocketConnector> {
             AtChopsUtil.generateSymmetricKey(EncryptionKeyType.aes256).key;
         String socketIV =
             base64Encode(AtChopsUtil.generateRandomIV(16).ivBytes);
-        controlSink.add(
-            Uint8List.fromList('connect:$socketAESKey:$socketIV\n'.codeUnits));
         // Authenticate the sideB socket (to the rvd)
-        if (rvdAuthString != null) {
-          logger
-              .info('_runClientSideMulti authenticating new connection to rvd');
-          sideB.socket.writeln(rvdAuthString);
+        try {
+          if (relayAuthenticator != null) {
+            logger.info('_runClientSideMulti  (_clientSideEncryptedSocket)'
+                ' authenticating new connection to rvd');
+            var (authenticated, authenticatedStream) =
+                await relayAuthenticator!.authenticate(sideB.socket);
+            if (!authenticated || authenticatedStream == null) {
+              sideB.socket.destroy();
+            } else {
+              sideB.stream = authenticatedStream;
+            }
+          }
+          sideA.transformer = createEncrypter(socketAESKey, socketIV);
+          sideB.transformer = createDecrypter(socketAESKey, socketIV);
+          logger.info('_runClientSideMulti (_clientSideEncryptedSocket)'
+              ' Client side connected.'
+              ' Sending connect request to daemon');
+          controlSink.add(Uint8List.fromList(
+              'connect:$socketAESKey:$socketIV\n'.codeUnits));
+        } catch (err) {
+          logger.severe('_runClientSideMulti (_clientSideEncryptedSocket)'
+              ' Failed to authenticate to relay: $err'
+              '\n\tWill destroy local socket also'); // TODO add retries like on daemon side
+          sideB.socket.destroy();
+          sideA.socket.destroy();
         }
-        sideA.transformer = createEncrypter(socketAESKey, socketIV);
-        sideB.transformer = createDecrypter(socketAESKey, socketIV);
       },
     );
     return socketConnector;
@@ -692,44 +795,76 @@ class SrvImplDart implements Srv<SocketConnector> {
 
     // First, connect to the relay.
     // If it fails, return, as we can't do anything more
-    late Socket sideBSocket;
     late Side sideB;
     logger.info(
-        'socket_connector: Connecting side B (relay - $relayAddress:$streamingPort)');
+        '_runDaemonSideMulti: Connecting side B (relay - $relayAddress:$streamingPort)');
+    bool candidateConnected = false;
     try {
-      sideBSocket = await Socket.connect(relayAddress, streamingPort);
-      sideB = Side(sideBSocket, false, transformer: decrypter);
-      unawaited(sideBSocket.done
-          .then((v) => logger.info('relay socket done'))
-          .catchError(
-              (err) => logger.warning('relay socket done with error $err')));
-      if (rvdAuthString != null) {
-        logger.info('_runDaemonSideMulti authenticating'
-            ' new socket connection to relay');
-        sideBSocket.writeln(rvdAuthString);
-        await sideBSocket.flush();
+      int attempts = 4;
+      for (int i = 1; i <= attempts && !candidateConnected; i++) {
+        Side candidateSideB = Side(
+            await Socket.connect(relayAddress, streamingPort), false,
+            transformer: decrypter);
+        unawaited(candidateSideB.socket.done
+            .then((v) => logger.info('relay socket done'))
+            .catchError(
+                (err) => logger.warning('relay socket done with error $err')));
+        if (relayAuthenticator == null) {
+          candidateConnected = true;
+        } else {
+          logger.info('_runDaemonSideMulti authenticating'
+              ' new socket connection to relay');
+          try {
+            var (authenticated, authenticatedStream) =
+                await relayAuthenticator!.authenticate(candidateSideB.socket);
+            logger.info('_runDaemonSideMulti authentication apparently complete'
+                '\n\t=> authenticated: $authenticated'
+                '\n\t=> stream: $authenticatedStream');
+            if (!authenticated || authenticatedStream == null) {
+              candidateSideB.socket.destroy();
+            } else {
+              candidateSideB.stream = authenticatedStream;
+              sideB = candidateSideB;
+              candidateConnected = true;
+            }
+          } catch (err) {
+            logger.info('_runDaemonSideMulti (_handleMultiConnectRequest)'
+                ' Failed to authenticate to relay: $err');
+            candidateSideB.socket.destroy();
+          }
+        }
+        if (!candidateConnected && i < attempts) {
+          logger.info('Will try again in 1 second');
+          await Future.delayed(Duration(seconds: 1));
+        }
+      }
+      if (!candidateConnected) {
+        logger.shout('Failed to authenticate to relay $attempts times');
+        sc.close();
+        return;
       }
     } catch (e) {
-      logger.shout('Failed to connect to relay ($relayAddress:$streamingPort)'
-          ' with error : $e');
+      logger.shout(
+          'Failed to connect to relay ($relayAddress:$streamingPort) with error : $e');
+      sc.close();
       return;
     }
 
     // Now, connect to the local host:port
     // If it fails, we need to close the socket connection to the relay
-    late Socket sideASocket;
+    late Socket sideASocketActual;
     late Side sideA;
-    logger.info(
-        'socket_connector: Connecting side A (local - $localAddress:$localPort)');
+    logger.info('_runDaemonSideMulti:'
+        ' Connecting side A (local - $localAddress:$localPort)');
 
     try {
-      sideASocket = await Socket.connect(localAddress, localPort);
-      sideA = Side(sideASocket, true, transformer: encrypter);
+      sideASocketActual = await Socket.connect(localAddress, localPort);
+      sideA = Side(sideASocketActual, true, transformer: encrypter);
     } catch (e) {
       logger.shout('Failed to connect locally ($localAddress:$localPort)'
           ' with error : $e');
       logger.shout('Closing sideB (relay) socket connection');
-      sideBSocket.destroy();
+      sideB.socket.destroy();
 
       return;
     }
@@ -737,11 +872,11 @@ class SrvImplDart implements Srv<SocketConnector> {
     // Finally, let's have the socket connector join the sides together
     unawaited(sc.handleSingleConnection(sideB).catchError((err) {
       logger.severe(
-          'ERROR $err from handleSingleConnection on sideB (relay - $relayAddress:$streamingPort)');
+          'ERROR $err in _runDaemonSideMulti from handleSingleConnection on sideB (relay - $relayAddress:$streamingPort)');
     }));
     unawaited(sc.handleSingleConnection(sideA).catchError((err) {
       logger.severe(
-          'ERROR $err from handleSingleConnection on sideA (local - $localAddress:$localPort)');
+          'ERROR $err in _runDaemonSideMulti from handleSingleConnection on sideA (local - $localAddress:$localPort)');
     }));
 
     logger.info('socket_connector: started');
@@ -788,6 +923,8 @@ class SrvImplDart implements Srv<SocketConnector> {
     required InternetAddress relayAddress,
     required Duration timeout,
   }) async {
+    logger.info(
+        '_runDaemonSideMulti: creating SocketConnector with timeout $timeout');
     SocketConnector sc = SocketConnector(
       timeout: timeout,
       verbose: Platform.environment['SRV_TRACE'] == 'true',
@@ -800,20 +937,47 @@ class SrvImplDart implements Srv<SocketConnector> {
         streamingHost, streamingPort,
         timeout: Duration(seconds: 10));
     // Authenticate the control channel
-    if (rvdAuthString != null) {
+    Stream<Uint8List>? authenticatedControlSocketStream;
+    if (relayAuthenticator != null) {
       logger.info('_runDaemonSideMulti authenticating'
           ' control channel connection to rvd');
-      sessionControlSocket.writeln(rvdAuthString);
+      bool authenticated;
+      try {
+        (authenticated, authenticatedControlSocketStream) =
+            await relayAuthenticator!.authenticate(sessionControlSocket);
+        if (!authenticated || authenticatedControlSocketStream == null) {
+          sessionControlSocket.destroy();
+          throw Exception('_runDaemonSideMulti'
+              ' Failed to authenticate control socket');
+        }
+      } catch (err) {
+        logger.severe('_runDaemonSideMulti'
+            ' Failed to authenticate control socket: $err');
+        sessionControlSocket.destroy();
+        rethrow;
+      }
+    } else {
+      authenticatedControlSocketStream = sessionControlSocket;
     }
 
     if (sessionAESKeyString != null && sessionIVString != null) {
       logger
           .info('_runDaemonSideMulti: On the daemon side traffic is encrypted');
-      _daemonSideEncryptedSocket(sessionControlSocket, sc, relayAddress);
+      _daemonSideEncryptedSocket(
+        authenticatedControlSocketStream,
+        sessionControlSocket,
+        sc,
+        relayAddress,
+      );
     } else {
       logger.info(
           '_runDaemonSideMulti: On the daemon side traffic is transmitted in plain text');
-      _daemonSidePlainSocket(sessionControlSocket, sc, relayAddress);
+      _daemonSidePlainSocket(
+        authenticatedControlSocketStream,
+        sessionControlSocket,
+        sc,
+        relayAddress,
+      );
     }
 
     // upon socketConnector.done, destroy the control channel, and complete
@@ -824,21 +988,26 @@ class SrvImplDart implements Srv<SocketConnector> {
     return sc;
   }
 
-  /// Start the control stream listener on the control channel itself
-  /// since there is no encryption / decryption
-  void _daemonSidePlainSocket(Socket sessionControlSocket, SocketConnector sc,
-      InternetAddress relayAddress) {
+  void _daemonSidePlainSocket(
+    Stream<Uint8List> sessionControlSocketStream,
+    IOSink sessionControlSocketSink,
+    SocketConnector sc,
+    InternetAddress relayAddress,
+  ) {
     _startDaemonControlSocketListener(
-      sessionControlSocket,
-      sessionControlSocket,
+      sessionControlSocketStream,
+      sessionControlSocketSink,
       sc,
       relayAddress,
     );
   }
 
-  /// Set up encrypter & decrypter and then start the control stream listener
-  void _daemonSideEncryptedSocket(Socket sessionControlSocket,
-      SocketConnector sc, InternetAddress relayAddress) {
+  void _daemonSideEncryptedSocket(
+    Stream<Uint8List> sessionControlSocketStream,
+    IOSink sessionControlSocketSink,
+    SocketConnector sc,
+    InternetAddress relayAddress,
+  ) {
     DataTransformer controlEncrypter =
         createEncrypter(sessionAESKeyString!, sessionIVString!);
     DataTransformer controlDecrypter =
@@ -846,9 +1015,10 @@ class SrvImplDart implements Srv<SocketConnector> {
 
     // Listen to stream which is decrypting the socket stream
     // Write to a stream controller which encrypts and writes to the socket
-    Stream<List<int>> controlStream = controlDecrypter(sessionControlSocket);
+    Stream<List<int>> controlStream =
+        controlDecrypter(sessionControlSocketStream);
     StreamController<Uint8List> controlSink = StreamController<Uint8List>();
-    controlEncrypter(controlSink.stream).listen(sessionControlSocket.add);
+    controlEncrypter(controlSink.stream).listen(sessionControlSocketSink.add);
 
     _startDaemonControlSocketListener(
       controlStream,
@@ -965,18 +1135,56 @@ class SrvImplDart implements Srv<SocketConnector> {
     }
     InternetAddress localAddress = await resolveRequestedLocalHost();
 
-    SocketConnector socketConnector = await SocketConnector.socketToSocket(
-        addressA: localAddress,
-        portA: localPort,
-        addressB: relayAddress,
-        portB: streamingPort,
-        verbose: Platform.environment['SRV_TRACE'] == 'true',
-        logger: ioSinkForLogger(logger),
-        transformAtoB: encrypter,
-        transformBtoA: decrypter);
-    if (rvdAuthString != null) {
+    bool verbose = Platform.environment['SRV_TRACE'] == 'true';
+    IOSink logSink = ioSinkForLogger(logger);
+    SocketConnector socketConnector = SocketConnector(
+      verbose: verbose,
+      logger: ioSinkForLogger(logger),
+    );
+    if (verbose) {
+      logSink.writeln('socket_connector:'
+          ' Connecting to $localAddress:$localPort');
+    }
+    Socket sideASocket = await Socket.connect(localAddress, localPort);
+    Side sideA = Side(sideASocket, true, transformer: encrypter);
+    unawaited(socketConnector.handleSingleConnection(sideA).catchError((err) {
+      logSink.writeln(
+          'ERROR $err in _runDaemonSideSingle from handleSingleConnection on sideA (local)');
+    }));
+
+    if (verbose) {
+      logSink.writeln(
+          'socket_connector: Connecting to $relayAddress:$streamingPort');
+    }
+    Socket sideBSocket = await Socket.connect(relayAddress, streamingPort);
+    Side sideB = Side(sideBSocket, false, transformer: decrypter);
+
+    // Authenticate the sideB socket (to the rvd)
+    if (relayAuthenticator != null) {
       logger.info('_runDaemonSideSingle authenticating socketB to rvd');
-      socketConnector.connections.first.sideB.socket.writeln(rvdAuthString);
+      try {
+        var (authenticated, authenticatedStream) =
+            await relayAuthenticator!.authenticate(sideB.socket);
+        if (!authenticated || authenticatedStream == null) {
+          sideB.socket.destroy();
+        } else {
+          sideB.stream = authenticatedStream;
+        }
+      } catch (err) {
+        logger.severe('_runDaemonSideSingle'
+            ' Failed to authenticate socketB: $err');
+        sideB.socket.destroy();
+        sideA.socket.destroy();
+      }
+    }
+
+    unawaited(socketConnector.handleSingleConnection(sideB).catchError((err) {
+      logSink.writeln(
+          'ERROR $err in _runDaemonSideSingle from handleSingleConnection on sideB (relay)');
+    }));
+
+    if (verbose) {
+      logSink.writeln('socket_connector: started');
     }
 
     return socketConnector;