+++ packages/dart/noports_core/lib/src/srvd/srvd.dart
@@ -1,12 +1,15 @@
 import 'dart:async';
+import 'dart:isolate';
 
 import 'package:at_client/at_client.dart';
 import 'package:at_utils/at_logger.dart';
 import 'package:meta/meta.dart';
+import 'package:noports_core/src/srvd/isolates/types.dart';
 import 'package:noports_core/src/srvd/srvd_impl.dart';
 import 'package:noports_core/src/srvd/srvd_params.dart';
+import 'package:noports_core/src/srvd/srvd_session_params.dart';
 
-abstract class Srvd {
+abstract interface class Srvd {
   static const String namespace = 'sshrvd';
 
   abstract final AtSignLogger logger;
@@ -17,11 +20,13 @@ abstract class Srvd {
   abstract final String managerAtsign;
   abstract final String ipAddress;
   abstract final bool logTraffic;
-  bool verbose = false;
+  abstract final bool bind443;
+  abstract final int localBindPort443;
+  abstract bool verbose;
 
   /// true once [init] has completed
   @visibleForTesting
-  bool initialized = false;
+  abstract bool initialized;
 
   static Future<Srvd> fromCommandLineArgs(List<String> args,
       {AtClient? atClient,
@@ -34,5 +39,18 @@ abstract class Srvd {
   }
 
   Future<void> init();
+
   Future<void> run();
+
+  Future<void> lookup(IIRequest msg, SendPort toSpawned);
+
+  Future<(PortPair, Isolate, SendPort)> spawnNewPortPairIsolate(
+    SrvdSessionParams sessionParams,
+  );
+
+  Future<(PortPair, Isolate, SendPort)> spawnNewSinglePortIsolate(
+    String address,
+    bool useTLS,
+    int bindPort,
+  );
 }