+++ packages/dart/noports_core/lib/src/sshnp/util/srvd_channel/notification_request_message.dart
@@ -1,12 +1,31 @@
 import 'dart:convert';
 
+import 'package:noports_core/sshnp_foundation.dart';
+
 class SocketRendezvousRequestMessage {
-  late String sessionId;
-  late String atSignA;
-  late String atSignB;
-  late bool authenticateSocketA;
-  late bool authenticateSocketB;
-  late String clientNonce;
+  final String sessionId;
+  final String atSignA;
+  final String atSignB;
+  final bool authenticateSocketA;
+  final bool authenticateSocketB;
+  final String clientNonce;
+  final RelayAuthMode relayAuthMode;
+  final String? relayAuthAesKey;
+  final bool only443;
+  final bool sendNacks;
+
+  SocketRendezvousRequestMessage({
+    required this.sessionId,
+    required this.atSignA,
+    required this.atSignB,
+    required this.authenticateSocketA,
+    required this.authenticateSocketB,
+    required this.clientNonce,
+    required this.relayAuthMode,
+    required this.relayAuthAesKey,
+    required this.only443,
+    required this.sendNacks,
+  });
 
   @override
   String toString() {
@@ -17,6 +36,10 @@ class SocketRendezvousRequestMessage {
     m['authenticateSocketA'] = authenticateSocketA;
     m['authenticateSocketB'] = authenticateSocketB;
     m['clientNonce'] = clientNonce;
+    m['relayAuthMode'] = relayAuthMode.name;
+    m['relayAuthAesKey'] = relayAuthAesKey;
+    m['only443'] = only443;
+    m['sendNacks'] = sendNacks;
     return jsonEncode(m);
   }
 }