+++ packages/dart/noports_core/lib/src/sshnpd/sshnpd_impl.dart
@@ -7,11 +7,14 @@ import 'package:at_client/at_client.dart' hide StringBuffer;
 import 'package:at_client/at_client_mixins.dart';
 import 'package:at_utils/at_logger.dart';
 import 'package:dartssh2/dartssh2.dart';
+import 'package:file/local.dart';
 import 'package:logging/logging.dart';
 import 'package:meta/meta.dart';
 import 'package:noports_core/src/common/features.dart';
 import 'package:noports_core/src/common/handle_server_events.dart';
+import 'package:noports_core/src/common/mixins/apkam_signing.dart';
 import 'package:noports_core/src/common/openssh_binary_path.dart';
+import 'package:noports_core/src/srv/relay_authenticators.dart';
 import 'package:noports_core/src/srv/srv.dart';
 import 'package:noports_core/src/sshnp/impl/notification_request_message.dart';
 import 'package:noports_core/sshnpd.dart';
@@ -22,7 +25,7 @@ import 'package:socket_connector/socket_connector.dart';
 import 'package:uuid/uuid.dart';
 
 @protected
-class SshnpdImpl with AtClientBindings implements Sshnpd {
+class SshnpdImpl with AtClientBindings, ApkamSigning implements Sshnpd {
   @override
   final AtSignLogger logger = AtSignLogger(' sshnpd ');
 
@@ -141,7 +144,9 @@ class SshnpdImpl with AtClientBindings implements Sshnpd {
         DaemonFeature.supportsPortChoice.name: true,
         DaemonFeature.adjustableTimeout.name: true,
         DaemonFeature.controlChannelHeartbeats.name: true,
+        DaemonFeature.supportsRamEscr.name: true,
       },
+      'authModes': RelayAuthMode.values.map((c) => c.name).toList(),
       'allowedServices': permitOpen,
       'npCpVersion': DaemonFeature.latestVersion.toString(),
     };
@@ -167,7 +172,7 @@ class SshnpdImpl with AtClientBindings implements Sshnpd {
         throw ('\n Unable to find .atKeys file : ${p.atKeysFilePath}');
       }
 
-      AtSignLogger.root_level = 'SHOUT';
+      AtSignLogger.root_level = 'SEVERE';
       if (p.verbose) {
         AtSignLogger.root_level = 'INFO';
       }
@@ -201,6 +206,14 @@ class SshnpdImpl with AtClientBindings implements Sshnpd {
         sshnpd.logger.logger.level = Level.INFO;
       }
 
+      if (p.clearCachedPKs) {
+        sshnpd.logger.shout('Clearing cached public keys');
+        await clearLocallyCachedPKs(
+          logger: sshnpd.logger,
+          fs: LocalFileSystem(),
+          atClient: sshnpd.atClient,
+        );
+      }
       return sshnpd;
     } catch (e, s) {
       usageCallback?.call(e, s);
@@ -213,6 +226,8 @@ class SshnpdImpl with AtClientBindings implements Sshnpd {
     if (initialized) {
       throw StateError('Cannot init() - already initialized');
     }
+
+    await publishPublicSigningKey();
 
     initialized = true;
   }
@@ -559,14 +574,29 @@ class SshnpdImpl with AtClientBindings implements Sshnpd {
     //      e.g. making the rvdAuthString; generating AES and IV and encrypting them
 
     try {
-      String? rvdAuthString;
+      RelayAuthenticator? relayAuthenticator;
+
       if (req.authenticateToRvd) {
-        // TODO refactor duplicate code from startDirectSsh
-        rvdAuthString = signAndWrapAndJsonEncode(atClient, {
-          'sessionId': req.sessionId,
-          'clientNonce': req.clientNonce,
-          'rvdNonce': req.rvdNonce,
-        });
+        switch (req.relayAuthMode) {
+          case RelayAuthMode.payload:
+            relayAuthenticator =
+                RelayAuthenticatorLegacy(signAndWrapAndJsonEncode(atClient, {
+              'sessionId': req.sessionId,
+              'clientNonce': req.clientNonce,
+              'rvdNonce': req.rvdNonce,
+            }));
+            break;
+          case RelayAuthMode.escr:
+            relayAuthenticator = RelayAuthenticatorESCR(
+              sessionId: req.sessionId,
+              relayAuthAesKey: req.relayAuthAesKey!,
+              publicSigningKeyUri: publicSigningKeyUri,
+              publicSigningKey: publicSigningKey,
+              privateSigningKey: privateSigningKey,
+              isSideA: false,
+            );
+            break;
+        }
       }
 
       String? sessionAESKey, sessionAESKeyEncrypted;
@@ -610,7 +640,7 @@ class SshnpdImpl with AtClientBindings implements Sshnpd {
           localPort: req.requestedPort,
           bindLocalPort: false,
           localHost: req.requestedHost,
-          rvdAuthString: rvdAuthString,
+          relayAuthenticator: relayAuthenticator,
           sessionAESKeyString: sessionAESKey,
           sessionIVString: sessionIV,
           multi: true,
@@ -626,7 +656,7 @@ class SshnpdImpl with AtClientBindings implements Sshnpd {
           localPort: req.requestedPort,
           bindLocalPort: false,
           localHost: req.requestedHost,
-          rvdAuthString: rvdAuthString,
+          relayAuthenticator: relayAuthenticator,
           sessionAESKeyString: sessionAESKey,
           sessionIVString: sessionIV,
           multi: true,
@@ -655,7 +685,7 @@ class SshnpdImpl with AtClientBindings implements Sshnpd {
         atKey: _createResponseAtKey(
             requestingAtsign: requestingAtsign, sessionId: req.sessionId),
         value:
-            'Failed to start up the daemon side of the srv socket tunnel : $e',
+            'Failed to start up the daemon side of the relay socket tunnel : $e',
         sessionId: req.sessionId,
       );
     }
@@ -770,6 +800,8 @@ class SshnpdImpl with AtClientBindings implements Sshnpd {
       port: int.parse(port),
       privateKey: _privateKey,
       remoteForwardPort: int.parse(remoteForwardPort),
+      relayAuthMode: RelayAuthMode.payload,
+      relayAuthAesKey: null,
     );
 
     String requested = '$localSshdHost:$localSshdPort';
@@ -812,29 +844,39 @@ class SshnpdImpl with AtClientBindings implements Sshnpd {
     required String requestingAtsign,
     required SshnpSessionRequest req,
   }) async {
-    String sessionId = req.sessionId;
-    String host = req.host;
-    int port = req.port;
     bool? authenticateToRvd = req.authenticateToRvd;
-    String? clientNonce = req.clientNonce;
-    String? rvdNonce = req.rvdNonce;
     bool? encryptRvdTraffic = req.encryptRvdTraffic;
     String? clientEphemeralPK = req.clientEphemeralPK;
     String? clientEphemeralPKType = req.clientEphemeralPKType;
 
     logger.info(
-        'Setting up ports for direct ssh session using ${sshClient.name} ($sshClient) from: $requestingAtsign session: $sessionId');
+        'Setting up ports for direct ssh session using ${sshClient.name} ($sshClient) from: $requestingAtsign session: ${req.sessionId}');
 
     authenticateToRvd ??= false;
     encryptRvdTraffic ??= false;
     try {
-      String? rvdAuthString;
+      RelayAuthenticator? relayAuthenticator;
       if (authenticateToRvd) {
-        rvdAuthString = signAndWrapAndJsonEncode(atClient, {
-          'sessionId': sessionId,
-          'clientNonce': clientNonce,
-          'rvdNonce': rvdNonce,
-        });
+        switch (req.relayAuthMode) {
+          case RelayAuthMode.payload:
+            relayAuthenticator =
+                RelayAuthenticatorLegacy(signAndWrapAndJsonEncode(atClient, {
+              'sessionId': req.sessionId,
+              'clientNonce': req.clientNonce,
+              'rvdNonce': req.rvdNonce,
+            }));
+            break;
+          case RelayAuthMode.escr:
+            relayAuthenticator = RelayAuthenticatorESCR(
+              sessionId: req.sessionId,
+              relayAuthAesKey: req.relayAuthAesKey!,
+              publicSigningKeyUri: publicSigningKeyUri,
+              publicSigningKey: publicSigningKey,
+              privateSigningKey: privateSigningKey,
+              isSideA: false,
+            );
+            break;
+        }
       }
 
       String? sessionAESKey, sessionAESKeyEncrypted;
@@ -875,11 +917,11 @@ class SshnpdImpl with AtClientBindings implements Sshnpd {
       // Connect to rendezvous point using background process.
       // This program can then exit without causing an issue.
       Process rv = await Srv.exec(
-        host,
-        port,
+        req.host,
+        req.port,
         localPort: localSshdPort,
         bindLocalPort: false,
-        rvdAuthString: rvdAuthString,
+        relayAuthenticator: relayAuthenticator,
         sessionAESKeyString: sessionAESKey,
         sessionIVString: sessionIV,
         timeout: DefaultArgs.srvTimeout,
@@ -891,12 +933,12 @@ class SshnpdImpl with AtClientBindings implements Sshnpd {
       /// Generate the ephemeral key pair which the client will use for the
       /// initial tunnel ssh session
       AtSshKeyPair tunnelKeyPair = await keyUtil.generateKeyPair(
-          algorithm: sshAlgorithm, identifier: 'ephemeral_$sessionId');
+          algorithm: sshAlgorithm, identifier: 'ephemeral_${req.sessionId}');
 
       await keyUtil.authorizePublicKey(
         sshPublicKey: tunnelKeyPair.publicKeyContents,
         localSshdPort: localSshdPort,
-        sessionId: sessionId,
+        sessionId: req.sessionId,
         permissions: ephemeralPermissions,
       );
 
@@ -911,30 +953,30 @@ class SshnpdImpl with AtClientBindings implements Sshnpd {
       ///   ephemeral private key
       await _notify(
         atKey: _createResponseAtKey(
-            requestingAtsign: requestingAtsign, sessionId: sessionId),
+            requestingAtsign: requestingAtsign, sessionId: req.sessionId),
         value: signAndWrapAndJsonEncode(atClient, {
           'status': 'connected',
-          'sessionId': sessionId,
+          'sessionId': req.sessionId,
           'ephemeralPrivateKey': tunnelKeyPair.privateKeyContents,
           'sessionAESKey': sessionAESKeyEncrypted,
           'sessionIV': sessionIVEncrypted,
         }),
-        sessionId: sessionId,
+        sessionId: req.sessionId,
       );
 
       /// - start a timer to remove the ephemeral key from `authorized_keys`
       ///   after 15 seconds
       Timer(const Duration(seconds: 15),
-          () => keyUtil.deauthorizePublicKey(sessionId));
+          () => keyUtil.deauthorizePublicKey(req.sessionId));
     } catch (e) {
       logger.severe('startDirectSsh failed with unexpected error : $e');
       // Notify sshnp that this session is NOT connected
       await _notify(
         atKey: _createResponseAtKey(
-            requestingAtsign: requestingAtsign, sessionId: sessionId),
+            requestingAtsign: requestingAtsign, sessionId: req.sessionId),
         value:
-            'Failed to start up the daemon side of the srv socket tunnel : $e',
-        sessionId: sessionId,
+            'Failed to start up the daemon side of the relay socket tunnel : $e',
+        sessionId: req.sessionId,
       );
     }
   }
@@ -1409,6 +1451,8 @@ class _NPAAuthChecker implements AuthChecker, AtRpcCallbacks {
       domainNameSpace: 'auth_checks',
       allowList: {sshnpd.policyManagerAtsign!},
       callbacks: this,
+      isClient: true,
+      isServer: false,
     );
     rpc.start();
   }